+++ passes/checks.go
@@ -35,6 +35,7 @@ import (
 	"github.com/bflad/tfproviderlint/passes/S020"
 	"github.com/bflad/tfproviderlint/passes/S021"
 	"github.com/bflad/tfproviderlint/passes/S022"
+	"github.com/bflad/tfproviderlint/passes/S023"
 	"github.com/bflad/tfproviderlint/passes/V001"
 	"golang.org/x/tools/go/analysis"
 )
@@ -77,5 +78,6 @@ var AllChecks = []*analysis.Analyzer{
 	S020.Analyzer,
 	S021.Analyzer,
 	S022.Analyzer,
+	S023.Analyzer,
 	V001.Analyzer,
 }